+++ index.js
@@ -1,34 +1,12 @@
 var stylelint = require("stylelint")
-var sass = require("node-sass")
-var postcss = require("postcss")
+var primerUtilities = require("primer-utilities")
 
 var ruleName = "primer/selector-no-utility"
 var messages = stylelint.utils.ruleMessages(ruleName, {
-  rejected: "Avoid styling the utilty class",
+  rejected: function(classname) { return "Avoid styling the utilty class `" + classname + "`"  },
 })
 
-function buildUtilityClasses(options) {
-  var utilityClasses = []
-
-  if (options == null) {
-    options = {}
-  }
-
-  var sassresult = sass.renderSync({
-    data: options["data"] || "@import \"primer-utilities/index.scss\";",
-    includePaths: [ "node_modules/" ].concat(options["include-paths"])
-  })
-
-  // post css walking
-  var utilRoot = postcss.parse(sassresult.css.toString())
-  utilRoot.walkRules(function(rule) {
-    utilityClasses = utilityClasses.concat(rule.selectors)
-  })
-
-  return utilityClasses
-}
-
-module.exports = stylelint.createPlugin(ruleName, function(enabled, options) {
+module.exports = stylelint.createPlugin(ruleName, function(enabled) {
   return function(root, result) {
 
     var validOptions = stylelint.utils.validateOptions(result, ruleName, {
@@ -38,13 +16,7 @@ module.exports = stylelint.createPlugin(ruleName, function(enabled, options) {
 
     if (!validOptions) { return }
 
-    var utilityClasses = []
-
-    try {
-      utilityClasses = buildUtilityClasses(options)
-    } catch (e) {
-      // Editor plugin won't work, needs correct path
-    }
+    var utilityClasses = primerUtilities.cssstats.selectors.values
 
     if (utilityClasses.length == 0) {
       return
@@ -52,16 +24,19 @@ module.exports = stylelint.createPlugin(ruleName, function(enabled, options) {
 
     root.walkRules(function(rule) {
 
-      var ruleClasses = rule.selector.split(" ")
+      var ruleClasses = rule.selector.match(/\.[a-z\-_0-9]+/ig)
 
-      for (var i = 0; i < ruleClasses.length; i++) {
-        if (utilityClasses.indexOf(ruleClasses[i]) >= 0) {
-          stylelint.utils.report({
-            message: messages.rejected,
-            node: rule,
-            result: result,
-            ruleName: ruleName
-          })
+      if (ruleClasses != null) {
+        for (var i = 0; i < ruleClasses.length; i++) {
+          var ruleClass = ruleClasses[i]
+          if (utilityClasses.indexOf(ruleClass) >= 0) {
+            stylelint.utils.report({
+              message: messages.rejected(ruleClass),
+              node: rule,
+              result: result,
+              ruleName: ruleName
+            })
+          }
         }
       }
     })