+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # HEAD
 
+# 1.1.0
+
+  * Updated: Using updated primer-utilities module to pull classes in through exports
+  * Added: Class name to error message
+
 # 1.0.0
 
   * Added: New rule `selector-no-utility`